+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # docs
 
+## 3.0.0
+
+### Major Changes
+
+- Create a new package in @agencecinq/shopify
+
 ## 2.0.0
 
 ### Major Changes