+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @agencecinq/utils
 
+## 4.0.0
+
+### Major Changes
+
+- Create a new package in @agencecinq/shopify
+
 ## 3.0.0
 
 ### Major Changes